src/main/java/avito/converter/service/validator/ValidUrlCheckerImpl.java
package avito.converter.service.validator;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import java.io.IOException;
import java.net.HttpURLConnection;
import java.net.URL;

@Service
@RequiredArgsConstructor
@Slf4j
public class ValidUrlCheckerImpl implements ValidUrlChecker{


    @Override
    public boolean validateURl(URL url) throws IOException {
        HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
        int responseCode = urlConnection.getResponseCode();
        return responseCode != 404;
    }
}
